Updated Code with src version d3557e8ccc8d7f7fed328e654c824c97fec6aec2

+++ Source/src/eShopLegacyWebForms/Models/Infrastructure/CatalogDBInitializer.cs
@@ -14,11 +14,6 @@ namespace eShopLegacyWebForms.Models.Infrastructure
         private const string CatalogBrandHiLoSequenceScript = @"Models\Infrastructure\dbo.catalog_brand_hilo.Sequence.sql";
         private const string CatalogTypeHiLoSequenceScript = @"Models\Infrastructure\dbo.catalog_type_hilo.Sequence.sql";
 
-        private const string ProjectDirectory = @"C:\DEV\WebMAPDemos\src\WebForms\EShop\src\eShopLegacyWebForms";
-        private const string ProjectPictures = @"C:\DEV\WebMAPDemos\src\WebForms\EShop\src\eShopLegacyWebForms\Pics";
-
-
-
         private CatalogItemHiLoGenerator indexGenerator;
         public CatalogDBInitializer()
         {
@@ -95,19 +90,21 @@ namespace eShopLegacyWebForms.Models.Infrastructure
 
         private void ExecuteScript(CatalogDBContext context, string scriptFile)
         {
-            var scriptFilePath = Path.Combine(ProjectDirectory, scriptFile);
+            var scriptFilePath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, scriptFile);
             context.Database.ExecuteSqlCommand(File.ReadAllText(scriptFilePath));
         }
 
         private void AddCatalogItemPictures()
         {
-            var contentRootPath = ProjectDirectory;
-            DirectoryInfo picturePath = new DirectoryInfo(Path.Combine(ProjectPictures));
-            foreach (FileInfo file in picturePath.GetFiles())
+            var contentRootPath = AppDomain.CurrentDomain.BaseDirectory; 
+            DirectoryInfo picturePath = new DirectoryInfo(Path.Combine(contentRootPath,@"\Pics"));
+            if (Directory.Exists(picturePath.ToString()))
             {
-                file.Delete();
+                foreach (FileInfo file in picturePath.GetFiles())
+                {
+                    file.Delete();
+                }
             }
-
             string zipFileCatalogItemPictures = Path.Combine(contentRootPath, "Setup", "CatalogItems.zip");
             ZipFile.ExtractToDirectory(zipFileCatalogItemPictures, picturePath.ToString());
         }
